Move frequently used common class and pre-save receiver functions to root project - PublishStateOptions - publish_state_pre_save - slugify_pre_save
