+++ .vscode/settings.json
@@ -2,5 +2,8 @@
     "[python]": {
         "editor.defaultFormatter": "ms-python.black-formatter"
     },
-    "python.formatting.provider": "none"
+    "python.formatting.provider": "none",
+    "cSpell.words": [
+        "djangoflix"
+    ]
 }
