+++ src/videos/models.py
@@ -4,13 +4,11 @@ from django.db.models.query import QuerySet
 from django.utils import timezone
 from django.utils.text import slugify
 
+from djangoflix.db.models import PublishStateOptions
+from djangoflix.db.receivers import publish_state_pre_save, slugify_pre_save
+
 
 # Create your models here.
-class PublishStateOptions(models.TextChoices):
-    # This class defines the choices could be used in state field
-    # It shows like this: CONSTANT = 'DB_VALUE', 'USER_DISPLAY_VALUE'
-    PUBLISH = "PU", "Publish"
-    DRAFT = "DR", "Draft"
 
 
 class VideoQuerySet(models.QuerySet):
@@ -70,28 +68,6 @@ class VideoPublishedProxy(Video):
         verbose_name_plural = "Published Videos"
 
 
-# publish state pre-save signal receiver function
-def publish_state_pre_save(sender, instance, *args, **kwargs):
-    is_publish = instance.state == PublishStateOptions.PUBLISH
-    is_draft = instance.state == PublishStateOptions.DRAFT
-
-    if is_publish and instance.publish_timestamp is None:
-        # add publish_timestamp if is_publish is True and publish_timestamp field is empty
-        instance.publish_timestamp = timezone.now()
-    elif is_draft:
-        # remove publish_timestamp field when video state is DRAFT
-        instance.publish_timestamp = None
-
-
+# Change publish_timestamp and slug fields with pre_save signal
 pre_save.connect(publish_state_pre_save, sender=Video)
-
-
-# slug field pre-save signal receiver function
-def slugify_pre_save(sender, instance, *args, **kwargs):
-    title = instance.title
-    slug = instance.slug
-    if slug is None:
-        instance.slug = slugify(title)
-
-
 pre_save.connect(slugify_pre_save, Video)
